[server] Add trackEvent to support dashboard analytics
